+++ test/wraft_doc_web/controllers/collection_form_field_controller_test.exs
@@ -4,7 +4,7 @@ defmodule WraftDocWeb.Api.V1.CollectionFormFieldControllerTest do
   """
   use WraftDocWeb.ConnCase
   @moduletag :controller
-  alias WraftDoc.Documents.CollectionFormField
+  alias WraftDoc.CollectionForms.CollectionFormField
   alias WraftDoc.Repo
 
   import WraftDoc.Factory
